"update prompt.py to match the latest version of it. 😁–“”—…‍‏•​․‎ †’"

+++ src/faceapi/database/models/prompt.py
@@ -4,7 +4,8 @@ from faceapi.database.fields import CleanCharField, CleanTextField
 from .base import DbModel
 from faceapi.database import Database
 from faceapi.routers.models import PromptResponse
-from peewee import FloatField, IntegerField, IntegrityError, BigAutoField
+from peewee import FloatField, IntegerField, IntegrityError, DoubleField
+from peewee import 
 from corestring import split_with_quotes, string_hash
 from argparse import ArgumentParser
 from pydantic import BaseModel, validator
@@ -65,7 +66,7 @@ class Prompt(DbModel):
     width = IntegerField(null=True)
     height = IntegerField(null=True)
     strength = FloatField(null=True)
-    seed = BigAutoField(null=True)
+    seed = DoubleField(null=True)
 
     @classmethod
     def get_hash(cls, **kwds) -> Optional[str]:
